requirements.txt
astroid==2.9.3
black==22.1.0
click==8.0.3
Flask==2.0.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
python-dotenv==0.19.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.0.1
Werkzeug==2.0.3
wrapt==1.13.3
